Checkpoint: Split executor and runner
