+++ bout_runners/runner/runner.py
@@ -1,0 +1,137 @@
+"""Contains the BOUT runner class."""
+
+
+from bout_runners.database.database_utils import tables_created
+from bout_runners.database.database_creator import create_database
+from bout_runners.database.database_utils import get_db_path
+from bout_runners.database.database_utils import \
+    get_file_modification
+from bout_runners.database.database_utils import \
+    get_system_info
+from bout_runners.database.database_utils import \
+    extract_parameters_in_use
+
+
+class BoutRunner:
+    """
+    Executes the command for submitting a run.
+
+    FIXME: Add variables and attributes
+
+    FIXME: Add examples
+    """
+
+    def __init__(self,
+                 executor,
+                 bookkeeper):
+        """
+        FIXME
+
+        Parameters
+        ----------
+        FIXME
+        """
+        # Set member data
+        self.executor = executor
+        self.bookkeeper = bookkeeper
+
+        # Get database
+        db_path = get_db_path(project_path=self.bout_paths.project_path,
+                              database_root_path=database_root_path)
+
+    def store_data_from_run(self,
+                            runner,
+                            number_of_processors,
+                            nodes=1,
+                            processors_per_node=None):
+        """
+        Capture data from a run.
+
+        Parameters
+        ----------
+        runner : BoutRunner
+            The bout runner object
+        number_of_processors : int
+            The total number of processors
+        nodes : int
+            The total number of nodes used
+        processors_per_node : int
+            Number of processors per nodes.
+            If None, this will be set to
+            floor(number_of_processors/nodes)
+
+        Returns
+        -------
+        new_entry : bool
+            Returns True if this a new entry is made, False if not
+        """
+        new_entry = False
+
+        # Initiate the run_dict (will be filled with the ids)
+        run_dict = {'name': runner.bout_paths.bout_inp_dst_dir.name}
+
+        # Update the parameters
+        parameters_dict = \
+            extract_parameters_in_use(
+                runner.bout_paths.project_path,
+                runner.bout_paths.bout_inp_dst_dir,
+                runner.run_parameters.run_parameters_dict)
+
+        run_dict['parameters_id'] = \
+            self.bookkeeper._create_parameter_tables_entry(
+                parameters_dict)
+
+        # Update the file_modification
+        file_modification_dict = \
+            get_file_modification(runner.bout_paths.project_path,
+                                  runner.make.makefile_path,
+                                  runner.make.exec_name)
+        run_dict['file_modification_id'] = \
+            self.check_entry_existence('file_modification',
+                                       file_modification_dict)
+        if run_dict['file_modification_id'] is None:
+            run_dict['file_modification_id'] = \
+                self.create_entry('file_modification',
+                                  file_modification_dict)
+
+        # Update the split
+        split_dict = {'number_of_processors': number_of_processors,
+                      'nodes': nodes,
+                      'processors_per_nodes': processors_per_node}
+        run_dict['split_id'] = \
+            self.check_entry_existence('split', split_dict)
+        if run_dict['split_id'] is not None:
+            run_dict['split_id'] = \
+                self.create_entry('split', split_dict)
+
+        # Update the system info
+        system_info_dict = get_system_info()
+        run_dict['host_id'] = \
+            self.check_entry_existence('system_info', system_info_dict)
+        if run_dict['host_id'] is not None:
+            run_dict['host_id'] = \
+                self.create_entry('system_info', system_info_dict)
+
+        # Update the run
+        run_id = self.check_entry_existence('run', run_dict)
+        if run_id is not None:
+            run_dict['latest_status'] = 'submitted'
+            self.create_entry('run', run_dict)
+            new_entry = True
+
+        return new_entry
+
+    def run(self):
+
+        if not tables_created(self.database):
+            db_path = self.database.database_path.parent
+            create_database(
+                project_path=self.bout_paths.project_path,
+                database_root_path=db_path)
+
+        self.database.store_data_from_run(
+            self,
+            self.processor_split.number_of_processors)
+
+        if tables_created(self.database) and not settings_run:
+            self.database.update_status()
