--- bout_runners/runners/__init__.py
@@ -1,1 +1,0 @@
-"""Package containing runners modules."""
